test: cast the size_t lengths that were implicitly narrowed
Seven call sites passed _oidc_strlen()'s size_t straight into an int or
unsigned int parameter. gcc is quiet about it, but the clang leg's
-Wimplicit-int-conversion group includes -Wshorten-64-to-32, so make check
could not be built with the static-analysis flag set. Make the conversions
explicit, matching the (int)_oidc_strlen idiom already used in the file.

Also run make check on that leg, so the type checks cover the suite rather
than only the module sources - clang is the only compiler here that catches
this class, gcc has no equivalent flag. Kept in one commit because the guard
and the fix have to land together for the leg to stay green.

+++ test/test_jose.c
@@ -1198,8 +1198,8 @@ START_TEST(test_jose_legacy_jwt_sign_verify_rsa_and_hmac) {
 
 	/* HMAC roundtrip with the same payload */
 	const char *secret = "mysecretwithmorethan32characters";
-	jwk =
-	    oidc_jwk_create_symmetric_key(pool, NULL, (const unsigned char *)secret, _oidc_strlen(secret), FALSE, &err);
+	jwk = oidc_jwk_create_symmetric_key(pool, NULL, (const unsigned char *)secret,
+					    (unsigned int)_oidc_strlen(secret), FALSE, &err);
 	ck_assert_ptr_nonnull(jwk);
 	apr_hash_set(keys, "dummy", APR_HASH_KEY_STRING, jwk);
 	jwt->header.alg = apr_pstrdup(pool, "HS256");
